refactor: type 폴더 변경
